Rename py_package_version -> py_module_version
