+++ gprofiler/metadata/py_module_version.py
